requirements.txt
pyyaml>=6.0
faker>=18.0.0
click>=8.0.0
requests>=2.28.0
python-dateutil>=2.8.0
colorama>=0.4.6